Move VariableRegistry ir/program/{block.rs -> mod.rs}
