+++ ir/translation/mod.rs
@@ -11,8 +11,9 @@ use answer::variable::Variable;
 use crate::{
     pattern::variable_category::VariableCategory,
     program::{
-        block::{BlockContext, ParameterRegistry, VariableRegistry},
+        block::{BlockContext, ParameterRegistry},
         function::Reducer,
+        VariableRegistry,
     },
 };
